Lights can be changed with the mouse wheel.
Updated the design for the main menu and credtis (to match a bit the
design proposed by Tatan).
Updated the default button styles to match the design proposed by Tatán.
Increased the movement speed of the PC.
Refactored the StateMachine for the player so now the movement is
handled by the Move state instead of the State script.
--- Ready to work on testing a different approach to the use of the
lights ---