Creature mechanic and path following.
Player has to hold LMB pressed while pointing to the creature in order
to capture it.
Creatures can follow a Path2D and leave pawprints in the ground.
Improved the way the Light for the pawprints and the pawprints itself
behave. Now we're not relying on the zIndex of the Light but the item
and light layers.
A progress bar is shown over the creature to show the capture progress.
Once the player has captured a creature, the HUD shows a panel that in
the future will have the card of the captured creature.
--- Ready to work on win condition, new lights and inventory ---